Refactoring Routing component into the navigator
