+++ app/questionnaire/questionnaire_manager.py
@@ -1,10 +1,111 @@
 import bleach
 from app.templating.template_pre_processor import TemplatePreProcessor
 from app.submitter.converter import SubmitterConstants
+from app.utilities.factory import factory
 from flask import session
 from flask_login import current_user
 from app.submitter.submitter import SubmitterFactory
 from app import settings
+
+# TODO I really like this pattern however it is in the wrong place here.
+# the navigator should deal with user action and it should call out to the
+# routing engine to deal with any routing requirements  which current there
+# aren't any so this component should simply do nothing.
+# Therefore all this functionality needs to move to the navigator and then I
+# think we combine and refactor that functionality
+
+
+class QuestionnaireManagerException(Exception):
+    pass
+
+
+class UserAction(object):
+    '''
+    Using the chain of responsibility pattern hand off a user action down the chain
+    until one of the objects can handle it
+    '''
+    def __init__(self, schema):
+        self._schema = schema
+        self.next_action = None
+
+    def set_next_action(self, action):
+        self.next_action = action
+        return action
+
+    def process_action(self, action, current_location):
+        # work through the user actions to determine which one matches the request action
+        if action == self.get_action():
+            return self.provide_next_location(current_location)
+        else:
+            # perform any functionality necessary to complete that action
+            self.perform_action()
+            # request the next location
+            return self.next_action.process_action(action, current_location)
+
+    def get_action(self):
+        pass
+
+    def perform_action(self):
+        pass
+
+    def provide_next_location(self, current_location):
+        pass
+
+
+class StartQuestionnaire(UserAction):
+    def get_action(self):
+        return 'start_questionnaire'
+
+    def provide_next_location(self, current_location):
+        return self._schema.groups[0].blocks[0].id
+
+
+class SaveContinue(UserAction):
+    def get_action(self):
+        return 'save_continue'
+
+    def provide_next_location(self, current_location):
+        # We have already been validated so we only
+        # need to figure out where to go next
+        next_location = current_location
+        current_block = self._schema.get_item_by_id(current_location)
+        if current_block:
+            current_group = current_block.container
+
+            for index, block in enumerate(current_group.blocks):
+                if block.id == current_block.id:
+                    if index + 1 < len(current_group.blocks):
+                        # return the next block in this group
+                        return current_group.blocks[index + 1].id
+
+            for index, group in enumerate(self._schema.groups):
+                if group.id == current_group.id:
+                    if index + 1 < len(self._schema.groups):
+                        # return the first block in the next group
+                        return self._schema.groups[index + 1].blocks[0].id
+
+            # There are no more blocks or groups, go to summary
+            next_location = 'summary'
+            return next_location
+        else:
+            raise QuestionnaireManagerException('Cannot route: No current block')
+
+
+class SubmitAnswers(UserAction):
+    def get_action(self):
+        return 'submit_answers'
+
+    def provide_next_location(self, current_location):
+        return 'thank-you'
+
+    def perform_action(self):
+        answer_store = factory.create("answer-store")
+        answers = answer_store.get_answers()
+        submitter = SubmitterFactory.get_submitter()
+        submitted_at = submitter.send_answers(current_user, self._metadata, self._schema, answers)
+        # TODO I don't like this but until we sort out the landing/review/submission flow this is the easiest way
+        session[SubmitterConstants.SUBMITTED_AT_KEY] = submitted_at.strftime(settings.DISPLAY_DATETIME_FORMAT)
+        return
 
 
 class QuestionnaireManager(object):
@@ -17,8 +118,19 @@ class QuestionnaireManager(object):
         self._navigator = navigator
         self._navigation_history = navigation_history
         self._metadata = metadata
-
         self.pre_processor = TemplatePreProcessor(self._schema, self.answer_store, self._validation_store, self._navigator, self._metadata)
+        self.user_actions = self._build_user_action_chain()
+
+    def _build_user_action_chain(self):
+        # build the chain of responsibility
+        start_questionnaire = StartQuestionnaire(self._schema)
+        save_continue = SaveContinue(self._schema)
+        submit = SubmitAnswers(self._schema)
+        start_questionnaire.set_next_action(save_continue).set_next_action(submit)
+        return start_questionnaire
+
+    def get_next(self, current_location, user_action):
+        return self.user_actions.process_action(user_action, current_location)
 
     def process_incoming_answers(self, post_data):
         # process incoming post data
@@ -40,10 +152,7 @@ class QuestionnaireManager(object):
         if self._validator.validate(cleaned_user_answers):
 
             # do any routing
-            next_location = self._routing_engine.get_next(current_location, user_action)
-
-            if next_location == 'thank-you':
-                self.submit()
+            next_location = self.get_next(current_location, user_action)
 
             # go to that location
             self._navigator.go_to(next_location)
@@ -53,14 +162,6 @@ class QuestionnaireManager(object):
 
         # now return the location
         return self._navigator.get_current_location()
-
-    def submit(self):
-        answers = self.answer_store.get_answers()
-        submitter = SubmitterFactory.get_submitter()
-        submitted_at = submitter.send_answers(current_user, self._metadata, self._schema, answers)
-        # TODO I don't like this but until we sort out the landing/review/submission flow this is the easiest way
-        session[SubmitterConstants.SUBMITTED_AT_KEY] = submitted_at.strftime(settings.DISPLAY_DATETIME_FORMAT)
-        return
 
     def get_rendering_context(self):
         return self.pre_processor.build_view_data()
